Update leaderboard instances
Remove seconds class for formatting, this was an old way to add a unit to a metric cell.
Remove finished todo comment in metric column class.
Update leaderboard column to allow the header field to be set.
Implements the sub table for the leaderboard instances of a submission entry.
Add benchmark instance serializer to the serializers.
Create leaderboard instance entry data processing, so we sub package that for the leaderboard, making it more clear.
Update leaderboard entry, based on the new sub leaderboard instance entry.
Update leaderboard processing to reflect changes to leaderboard entry and additional check for if a submission is verified, otherwise we do not rank it.